refactored well bottom clearance out of cores
