+++ api/src/opentrons/protocol_api/config.py
@@ -1,0 +1,23 @@
+"""Contains configuration related classes for Protocol API."""
+
+
+class Clearances:
+    def __init__(self, default_aspirate: float, default_dispense: float) -> None:
+        self._aspirate = default_aspirate
+        self._dispense = default_dispense
+
+    @property
+    def aspirate(self) -> float:
+        return self._aspirate
+
+    @aspirate.setter
+    def aspirate(self, new_val: float) -> None:
+        self._aspirate = float(new_val)
+
+    @property
+    def dispense(self) -> float:
+        return self._dispense
+
+    @dispense.setter
+    def dispense(self, new_val: float) -> None:
+        self._dispense = float(new_val)
