+++ api/src/opentrons/protocol_api/core/protocol_api/instrument_context.py
@@ -11,7 +11,6 @@ from opentrons.protocols.api_support.definitions import MAX_SUPPORTED_VERSION
 from opentrons.protocols.api_support.labware_like import LabwareLike
 from opentrons.protocols.api_support.types import APIVersion
 from opentrons.protocols.api_support.util import (
-    Clearances,
     build_edges,
     FlowRates,
     PlungerSpeeds,
@@ -48,9 +47,6 @@ class InstrumentContextImplementation(AbstractInstrument[WellImplementation]):
         self._mount = mount
         self._instrument_name = instrument_name
         self._default_speed = default_speed
-        self._well_bottom_clearances = Clearances(
-            default_aspirate=1.0, default_dispense=1.0
-        )
         self._flow_rates = FlowRates(self)
         self._speeds = PlungerSpeeds(self)
         self._flow_rates.set_defaults(api_level=self._api_version)
@@ -407,10 +403,6 @@ class InstrumentContextImplementation(AbstractInstrument[WellImplementation]):
     def get_return_height(self) -> float:
         """The height to return a tip to its tiprack."""
         return self.get_hardware_state().get("return_tip_height", 0.5)
-
-    def get_well_bottom_clearance(self) -> Clearances:
-        """The distance above the bottom of a well to aspirate or dispense."""
-        return self._well_bottom_clearances
 
     def get_flow_rate(self) -> FlowRates:
         return self._flow_rates
