+++ api/src/opentrons/protocols/api_support/util.py
@@ -246,28 +246,6 @@ class PlungerSpeeds:
         )
 
 
-class Clearances:
-    def __init__(self, default_aspirate: float, default_dispense: float) -> None:
-        self._aspirate = default_aspirate
-        self._dispense = default_dispense
-
-    @property
-    def aspirate(self) -> float:
-        return self._aspirate
-
-    @aspirate.setter
-    def aspirate(self, new_val: float):
-        self._aspirate = float(new_val)
-
-    @property
-    def dispense(self) -> float:
-        return self._dispense
-
-    @dispense.setter
-    def dispense(self, new_val: float):
-        self._dispense = float(new_val)
-
-
 class AxisMaxSpeeds(UserDict):
     """Special mapping allowing internal storage by Mount enums and
     user access by string
